set relationships

+++ database/migrations/2020_06_25_150850_create_questions_table.php
@@ -18,8 +18,8 @@ class CreateQuestionsTable extends Migration
             $table->string('title');
             $table->string('slug');
             $table->text('body');
-            $table->integer('category_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->unsignedBigInteger('category_id')->unsigned();
+            $table->unsignedBigInteger('user_id')->unsigned();
             $table->timestamps();
         });
     }
