+++ database/migrations/2020_06_25_151231_create_replies_table.php
@@ -16,8 +16,9 @@ class CreateRepliesTable extends Migration
         Schema::create('replies', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->text('body');
-            $table->integer('question_id')->unsigned();
-            $table->integer('user_id');
+            $table->unsignedBigInteger('question_id')->unsigned();
+            $table->unsignedBigInteger('user_id');
+
             $table->foreign('question_id')->references('id')->on('questions')
                 ->onDelete('cascade');
             $table->timestamps();
